Action Charging
We've started implementing action charging into the game, having made it compatible with all the retrigger types, properly handle cancellation (Including allowing you to trigger it even if it isn't fully charged), and letting ActionEffects be affected by the charge percentage.

To-do: We haven't yet added charge retention (Currently it always resets to 0), and we've also noticed that our mech weapon GFXs aren't updating in the mech build scene.

+++ NGP-Coursework/Assets/!_MainProject/Code/Scripts/GameplayObjects/Health/NetworkHealthState.cs
@@ -8,6 +8,6 @@ namespace Gameplay.GameplayObjects.Health
     /// </summary>
     public class NetworkHealthState : NetworkBehaviour
     {
-        public NetworkVariable<int> CurrentHealth = new NetworkVariable<int>();
+        public NetworkVariable<float> CurrentHealth = new NetworkVariable<float>();
     }
 }
